farm/pyproject.toml
[project]
name = "farm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "passlib[bcrypt]>=1.7.4",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "reflex>=0.8.28.post1",
]
